backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
